[PLT-9] 키, 팔길이 double로 변경 (#175)

+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/user/domain/User.kt
@@ -6,7 +6,7 @@ data class User(
     var name: String? = null,
     val provider: Provider,
     var isDeleted: Boolean = false,
-    val height: Int? = null,
-    val armSpan: Int? = null,
+    val height: Double? = null,
+    val armSpan: Double? = null,
     val instagramUrl: String? = null,
 )
